disable xsendfile
xsendfile is eating large amounts of CPU for unknown reasons (perhaps it
is not actually using sendfile?) and i am unable to get
apache/xsendfilet to recompile for use under a profiler to find up
what's going on.  so, for now, do the inefficient thing.  there's still
minimal CPU use though and the webapp is properly multithreaded